WIP: Refactoring USB protocol
